// frontend/src/app/api/models/history-series.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface HistorySeries {
  avg: Array<(number | null)>;
  max: Array<(number | null)>;
  metric: 'gpu_util' | 'vram_used_mib' | 'temp_c' | 'power_w' | 'cpu_percent' | 'ram_percent';
  scaleMax: (number | null);
  unit: string;
  warn: (number | null);
}